requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arrow==1.2.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
binaryornot==0.4.4
boto3==1.26.102
botocore==1.29.102
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
decorator==5.1.1
Django==4.1.4
django-autoslug==1.9.8
django-browser-reload==1.6.0
django-ckeditor==6.5.1
django-environ==0.10.0
django-extensions==3.2.1
django-js-asset==2.0.0
django-leaflet==0.28.3
django-leaflet-storage==0.8.2
django-social-share==2.3.0
django-storages==1.13.2
django-tailwind==3.4.0
frozenlist==1.3.3
future==0.18.3
geocoder==1.38.1
googlemaps==4.10.0
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
MarkupSafe==2.1.1
multidict==6.0.4
openai==0.27.0
Pillow==9.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==7.0.0
PyYAML==6.0
ratelim==0.1.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tqdm==4.65.0
urllib3==1.26.13
whitenoise==6.2.0
yarl==1.8.2
